+++ wxt.config.ts
@@ -22,7 +22,7 @@ export default defineConfig({
     name: 'Alfred - Shopify Theme Detector',
     description:
       'Instantly detect themes on any Shopify store. Streamline your workflow with smart shortcuts and Shopify productivity tools.',
-    version: '2026.08.06',
+    version: '2026.08.07',
     action: {
       default_title: 'Alfred'
     },